requirements.txt
attr==0.3.2
colorama==0.4.6
ConfigParser==7.1.0
cryptography==44.0.1
docutils==0.21.2
filelock==3.17.0
fpdf==1.7.2
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==8.1.5
jnius==1.1.0
keyring==25.6.0
llama_index==0.12.19
nltk==3.9.1
pandas==2.2.3
Pillow==11.1.0
protobuf==5.29.3
pyOpenSSL==25.0.0
python-dotenv==1.0.1
PyYAML==6.0.2
PyYAML==6.0.2
redis==5.2.1
Sphinx==8.2.0
streamlit==1.40.0
thread==2.0.5
transformers==4.48.0
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
